+++ src/lantern/exporters/cloudfront.py
@@ -3,6 +3,7 @@ from collections.abc import Collection
 from uuid import uuid4
 
 from mypy_boto3_cloudfront import CloudFrontClient
+from mypy_boto3_cloudfront.type_defs import InvalidationBatchTypeDef
 
 from lantern.exporters.base import ExporterBase
 from lantern.models.site import SiteContent
@@ -16,21 +17,18 @@ class CloudFrontExporter(ExporterBase):
         self._cf = cloudfront
         self._distribution = distribution
 
+        if not self._distribution:
+            msg = "No distribution specified"
+            raise ValueError(msg) from None
+
     def _invalidate_keys(self, keys: list[str]) -> None:
         """Create and execute CloudFront invalidation for selected keys."""
         caller_ref = str(uuid4())
         self._logger.info(f"Creating CloudFront invalidation for distribution {self._distribution}")
         self._logger.debug(keys)
-        response = self._cf.create_invalidation(
-            DistributionId=self._distribution,
-            InvalidationBatch={
-                "Paths": {
-                    "Quantity": len(keys),
-                    "Items": keys,
-                },
-                "CallerReference": caller_ref,
-            },
-        )
+
+        job: InvalidationBatchTypeDef = {"CallerReference": caller_ref, "Paths": {"Quantity": len(keys), "Items": keys}}
+        response = self._cf.create_invalidation(DistributionId=self._distribution, InvalidationBatch=job)
         invalidation_id = response["Invalidation"]["Id"]
         self._logger.info(f"Invalidation created: {invalidation_id}")
 
@@ -53,7 +51,7 @@ class CloudFrontExporter(ExporterBase):
 
         batches: list[list[str]] = [keys[i : i + 150] for i in range(0, len(keys), 150)]
         batch_count = len(batches)
-        self._logger.info(f"Invalidating {len(keys)} in {batch_count} 150-max key batches")
+        self._logger.info(f"Invalidating {len(keys)} keys in {batch_count} 150-max key batches")
         for i, batch in enumerate(batches):
             self._logger.info(f"Invalidating batch {i + 1} of {batch_count}")
             self._invalidate_keys(batch)